split engine and purge manager
